requirements.txt
anthropic==0.94.0
pandas==3.0.2
python-dotenv==1.2.2
streamlit==1.56.0
